Consolidate macro in one single file
